added search project assignment api
