+++ src/repositories/user-repository.ts
@@ -44,3 +44,9 @@ export const getAllByFilters = async (where: Prisma.usersWhereInput) => {
     where,
   })
 }
+
+export const getAllByFilters = async (where: Prisma.usersWhereInput) => {
+  return await prisma.users.findMany({
+    where,
+  })
+}
