+++ src/routes/admin/project-member-route.ts
@@ -1,14 +1,27 @@
-import express from "express"
-import * as ProjectMemberController from "../../controllers/admin/project-member-controller"
+import prisma from "../utils/prisma"
 
-const router = express.Router()
+export const getById = async (id: number) => {
+  return await prisma.projects.findUnique({
+    select: {
+      id: true,
+      name: true,
+    },
+    where: {
+      id,
+    },
+  })
+}
 
-router.get("/search", ProjectMemberController.search)
-
-router.get("/", ProjectMemberController.index)
-router.post("/", ProjectMemberController.store)
-router.get("/:id", ProjectMemberController.show)
-router.patch("/:id", ProjectMemberController.update)
-router.delete("/:id", ProjectMemberController.destroy)
-
-export default router
+export const getAllByName = async (name: string) => {
+  return await prisma.projects.findMany({
+    select: {
+      id: true,
+      name: true,
+    },
+    where: {
+      name: {
+        contains: name,
+      },
+    },
+  })
+}
